Fixes several protocol tests
