+++ smithy-swift-codegen/src/main/kotlin/software/amazon/smithy/swift/codegen/integration/serde/DynamicNodeDecodingGeneratorStrategy.kt
@@ -5,8 +5,11 @@
 
 package software.amazon.smithy.swift.codegen.integration.serde
 
+import software.amazon.smithy.aws.traits.protocols.RestXmlTrait
 import software.amazon.smithy.model.shapes.Shape
+import software.amazon.smithy.model.traits.XmlAttributeTrait
 import software.amazon.smithy.swift.codegen.integration.ProtocolGenerator
+import software.amazon.smithy.swift.codegen.integration.isInHttpBody
 import software.amazon.smithy.swift.codegen.integration.serde.xml.DynamicNodeDecodingXMLGenerator
 
 class DynamicNodeDecodingGeneratorStrategy(
@@ -23,4 +26,15 @@ class DynamicNodeDecodingGeneratorStrategy(
     private fun shouldRenderDynamicNodeDecodingProtocol(ctx: ProtocolGenerator.GenerationContext, shape: Shape): Boolean {
         return isRestXmlProtocolAndHasXmlAttributesInMembers(ctx, shape)
     }
+}
+
+fun isRestXmlProtocolAndHasXmlAttributesInMembers(ctx: ProtocolGenerator.GenerationContext, shape: Shape): Boolean {
+    val isRestXML = ctx.protocol == RestXmlTrait.ID
+    if (isRestXML) {
+        return shape.members()
+            .filter { it.isInHttpBody() }
+            .filter { it.hasTrait(XmlAttributeTrait::class.java) }
+            .isNotEmpty()
+    }
+    return false
 }
