+++ smithy-swift-codegen/src/main/kotlin/software/amazon/smithy/swift/codegen/integration/serde/xml/StructEncodeXMLGenerator.kt
@@ -5,6 +5,7 @@
 
 package software.amazon.smithy.swift.codegen.integration.serde.json
 
+import software.amazon.smithy.model.shapes.BlobShape
 import software.amazon.smithy.model.shapes.CollectionShape
 import software.amazon.smithy.model.shapes.MapShape
 import software.amazon.smithy.model.shapes.MemberShape
@@ -89,6 +90,9 @@ class StructEncodeXMLGenerator(
                 } else {
                     renderTimestampMember(member, memberTarget, containerName)
                 }
+            }
+            is BlobShape -> {
+                renderScalarMember(member, memberTarget, containerName, ".base64EncodedString()")
             }
             else -> {
                 renderScalarMember(member, memberTarget, containerName)
